fix bug where chunk spawn on each other
