--- Assets/ChunkDeath.cs
@@ -1,16 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class ChunkDeath : MonoBehaviour
-{
-    [SerializeField] private ChunkManager chunkManager;
-
-    private void OnTriggerEnter2D(Collider2D collision)
-    {
-        if (collision.CompareTag("Player"))
-        {
-            chunkManager.OnTouchBoundaries();
-        }
-    }
-}
