+++ Assets/EndlessMap/ChunkManager.prefab
@@ -11,7 +11,6 @@ GameObject:
   - component: {fileID: 3119443176499140269}
   - component: {fileID: 6164074325594727336}
   - component: {fileID: 5262753917999294219}
-  - component: {fileID: 7753174370120486298}
   m_Layer: 0
   m_Name: ChunkTopGeneration
   m_TagString: Untagged
@@ -93,20 +92,6 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   chunkManager: {fileID: 2877401654065343166}
   isTop: 1
---- !u!114 &7753174370120486298
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2781671411162406913}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 64368cb0ce5b38147bc340e322aec43f, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  chunkManager: {fileID: 3389069093796325714}
-  isTop: 1
 --- !u!1 &2904284887094598382
 GameObject:
   m_ObjectHideFlags: 0
@@ -151,8 +136,7 @@ GameObject:
   m_Component:
   - component: {fileID: 7793328810937219532}
   - component: {fileID: 4677696391368932576}
-  - component: {fileID: 1774080221126728361}
-  - component: {fileID: 5459030376089482955}
+  - component: {fileID: 9052719158400590205}
   m_Layer: 0
   m_Name: ChunkBottomGeneration
   m_TagString: Untagged
@@ -220,7 +204,7 @@ BoxCollider2D:
   serializedVersion: 2
   m_Size: {x: 10, y: 0.2}
   m_EdgeRadius: 0
---- !u!114 &1774080221126728361
+--- !u!114 &9052719158400590205
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -234,20 +218,6 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   chunkManager: {fileID: 2877401654065343166}
   isTop: 0
---- !u!114 &5459030376089482955
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4248993188140275463}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 64368cb0ce5b38147bc340e322aec43f, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  chunkManager: {fileID: 3389069093796325714}
-  isTop: 0
 --- !u!1 &7303827876546267535
 GameObject:
   m_ObjectHideFlags: 0
@@ -259,8 +229,7 @@ GameObject:
   - component: {fileID: 8836199867191011309}
   - component: {fileID: 1034222920357937841}
   - component: {fileID: 4845326584921333790}
-  - component: {fileID: 7321854655164097263}
-  - component: {fileID: 4886265375875891455}
+  - component: {fileID: 1138265128542648742}
   m_Layer: 0
   m_Name: ChunkDeath
   m_TagString: Untagged
@@ -318,10 +287,10 @@ BoxCollider2D:
   m_Offset: {x: -5, y: 0}
   m_SpriteTilingProperty:
     border: {x: 0, y: 0, z: 0, w: 0}
-    pivot: {x: 0, y: 0}
-    oldSize: {x: 0, y: 0}
-    newSize: {x: 0, y: 0}
-    adaptiveTilingThreshold: 0
+    pivot: {x: 0.5, y: 0.5}
+    oldSize: {x: 1, y: 1}
+    newSize: {x: 1, y: 1}
+    adaptiveTilingThreshold: 0.5
     drawMode: 0
     adaptiveTiling: 0
   m_AutoTiling: 0
@@ -363,17 +332,17 @@ BoxCollider2D:
   m_Offset: {x: 5, y: 0}
   m_SpriteTilingProperty:
     border: {x: 0, y: 0, z: 0, w: 0}
-    pivot: {x: 0, y: 0}
-    oldSize: {x: 0, y: 0}
-    newSize: {x: 0, y: 0}
-    adaptiveTilingThreshold: 0
+    pivot: {x: 0.5, y: 0.5}
+    oldSize: {x: 1, y: 1}
+    newSize: {x: 1, y: 1}
+    adaptiveTilingThreshold: 0.5
     drawMode: 0
     adaptiveTiling: 0
   m_AutoTiling: 0
   serializedVersion: 2
   m_Size: {x: 0.2, y: 10}
   m_EdgeRadius: 0
---- !u!114 &7321854655164097263
+--- !u!114 &1138265128542648742
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -385,20 +354,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 7d5ed2a2450a2434692e120bbbe3a0c6, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  chunkManager: {fileID: 0}
---- !u!114 &4886265375875891455
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7303827876546267535}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 7d5ed2a2450a2434692e120bbbe3a0c6, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  chunkManager: {fileID: 2877401654065343166}
+  _chunkManager: {fileID: 2877401654065343166}
 --- !u!1 &8385240334416708044
 GameObject:
   m_ObjectHideFlags: 0
@@ -409,9 +365,8 @@ GameObject:
   m_Component:
   - component: {fileID: 8519998686686959315}
   - component: {fileID: 2877401654065343166}
-  - component: {fileID: 3389069093796325714}
   m_Layer: 0
-  m_Name: Chunk
+  m_Name: ChunkManager
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -448,22 +403,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 602df2f4fece8bc48b44f477babd0a94, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  chunkOffset: 7
   chunk: {fileID: 8385240334416708044}
-  chunkOffset: 10
---- !u!114 &3389069093796325714
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8385240334416708044}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 602df2f4fece8bc48b44f477babd0a94, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  chunk: {fileID: 8385240334416708044}
-  chunkOffset: 0
 --- !u!1001 &221398561423441326
 PrefabInstance:
   m_ObjectHideFlags: 0
