+++ Assets/Magnets/Magnet.cs
@@ -69,10 +69,9 @@ public class Magnet : MonoBehaviour
 
     private void OnCollisionEnter2D(Collision2D other)
     {
-        if (other.gameObject.tag == "Player")
+        if (other.gameObject.CompareTag("Player"))
         {
-            FindObjectOfType<GameManager>().FailGame();
+            GameManager.Instance.FailGame();
         }
-        GameManager.Instance.FailGame();
     }
 }
