+++ Assets/Scenes/SampleScene.unity
@@ -1237,25 +1237,6 @@ Transform:
   - {fileID: 343748349}
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!1 &1796323081 stripped
-GameObject:
-  m_CorrespondingSourceObject: {fileID: 2781671411162406913, guid: 60cb5e2362bf9d344b3c5a5cad28e989, type: 3}
-  m_PrefabInstance: {fileID: 7361733865280533843}
-  m_PrefabAsset: {fileID: 0}
---- !u!114 &1796323084
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1796323081}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 64368cb0ce5b38147bc340e322aec43f, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  chunkManager: {fileID: 7361733865280533847}
-  isTop: 1
 --- !u!1 &1922679349
 GameObject:
   m_ObjectHideFlags: 0
@@ -1501,25 +1482,6 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1958632352}
   m_CullTransparentMesh: 1
---- !u!1 &1981847121 stripped
-GameObject:
-  m_CorrespondingSourceObject: {fileID: 4248993188140275463, guid: 60cb5e2362bf9d344b3c5a5cad28e989, type: 3}
-  m_PrefabInstance: {fileID: 7361733865280533843}
-  m_PrefabAsset: {fileID: 0}
---- !u!114 &1981847124
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1981847121}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 64368cb0ce5b38147bc340e322aec43f, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  chunkManager: {fileID: 7361733865280533847}
-  isTop: 0
 --- !u!1001 &1982539641
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -1721,7 +1683,7 @@ PrefabInstance:
     m_Modifications:
     - target: {fileID: 8385240334416708044, guid: 60cb5e2362bf9d344b3c5a5cad28e989, type: 3}
       propertyPath: m_Name
-      value: Chunk
+      value: ChunkManager
       objectReference: {fileID: 0}
     - target: {fileID: 8519998686686959315, guid: 60cb5e2362bf9d344b3c5a5cad28e989, type: 3}
       propertyPath: m_LocalPosition.x
@@ -1764,60 +1726,10 @@ PrefabInstance:
       value: 0
       objectReference: {fileID: 0}
     m_RemovedComponents: []
-    m_RemovedGameObjects:
-    - {fileID: 6759097502414212590, guid: 60cb5e2362bf9d344b3c5a5cad28e989, type: 3}
+    m_RemovedGameObjects: []
     m_AddedGameObjects: []
-    m_AddedComponents:
-    - targetCorrespondingSourceObject: {fileID: 8385240334416708044, guid: 60cb5e2362bf9d344b3c5a5cad28e989, type: 3}
-      insertIndex: -1
-      addedObject: {fileID: 7361733865280533847}
-    - targetCorrespondingSourceObject: {fileID: 7303827876546267535, guid: 60cb5e2362bf9d344b3c5a5cad28e989, type: 3}
-      insertIndex: -1
-      addedObject: {fileID: 7361733865280533848}
-    - targetCorrespondingSourceObject: {fileID: 2781671411162406913, guid: 60cb5e2362bf9d344b3c5a5cad28e989, type: 3}
-      insertIndex: -1
-      addedObject: {fileID: 1796323084}
-    - targetCorrespondingSourceObject: {fileID: 4248993188140275463, guid: 60cb5e2362bf9d344b3c5a5cad28e989, type: 3}
-      insertIndex: -1
-      addedObject: {fileID: 1981847124}
+    m_AddedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 60cb5e2362bf9d344b3c5a5cad28e989, type: 3}
---- !u!1 &7361733865280533845 stripped
-GameObject:
-  m_CorrespondingSourceObject: {fileID: 8385240334416708044, guid: 60cb5e2362bf9d344b3c5a5cad28e989, type: 3}
-  m_PrefabInstance: {fileID: 7361733865280533843}
-  m_PrefabAsset: {fileID: 0}
---- !u!1 &7361733865280533846 stripped
-GameObject:
-  m_CorrespondingSourceObject: {fileID: 7303827876546267535, guid: 60cb5e2362bf9d344b3c5a5cad28e989, type: 3}
-  m_PrefabInstance: {fileID: 7361733865280533843}
-  m_PrefabAsset: {fileID: 0}
---- !u!114 &7361733865280533847
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7361733865280533845}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 602df2f4fece8bc48b44f477babd0a94, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  chunk: {fileID: 0}
-  chunkOffset: 0
---- !u!114 &7361733865280533848
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7361733865280533846}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 7d5ed2a2450a2434692e120bbbe3a0c6, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  chunkManager: {fileID: 0}
 --- !u!1001 &8975987309164525339
 PrefabInstance:
   m_ObjectHideFlags: 0
